src/main/java/com/workingagile/acsd/BankAccount.java
package com.workingagile.acsd;

public class BankAccount {

    private int balance;

    public BankAccount(int initialBalance) {
        this.balance = initialBalance;
    }

    public int getBalance() {
        return balance;
    }

    public void deposit(int amount) {
        balance = balance + amount;
    }

    public void withdraw(int amount) throws InsufficientBalanceException {
        if (amount > balance) {
            throw new InsufficientBalanceException();
        }
        balance = balance - amount;
    }

    public static class InsufficientBalanceException extends Exception {}
}
